# src/test/resources/openapi.yaml
openapi: 3.0.1
info:
  title: RPC-REST Service API
  description: A Spring Boot application demonstrating REST and RPC endpoints
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Development server
- url: https://api.example.com
  description: Production server
tags:
- name: User Management
  description: REST API for managing users
- name: RPC Service
  description: Unified JSON-RPC endpoint for system operations
paths:
  /api/users/{id}:
    get:
      tags:
      - User Management
      summary: Get user by ID
      description: Retrieves a specific user by their unique identifier
      operationId: getUserById
      parameters:
      - name: id
        in: path
        description: User ID to retrieve
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: User not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponse'
        "200":
          description: User found successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponse'
    put:
      tags:
      - User Management
      summary: Update user
      description: Updates an existing user's information
      operationId: updateUser
      parameters:
      - name: id
        in: path
        description: User ID to update
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
        required: true
      responses:
        "404":
          description: User not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponse'
        "400":
          description: Invalid input data
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponse'
        "200":
          description: User updated successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponse'
    delete:
      tags:
      - User Management
      summary: Delete user
      description: Deletes a user from the system
      operationId: deleteUser
      parameters:
      - name: id
        in: path
        description: User ID to delete
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: User not found
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponse'
        "200":
          description: User deleted successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /rpc:
    post:
      tags:
      - RPC Service
      summary: Execute RPC call
      description: Unified JSON-RPC endpoint using discriminated union for params based on method field
      operationId: handleUnifiedRpcCall
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UnifiedRpcRequest'
            examples:
              Reboot with typed params:
                description: Reboot with typed params
                value:
                  method: reboot
                  params:
                    reason: Scheduled maintenance
                    force: false
                  id: "1"
              Status with typed params:
                description: Status with typed params
                value:
                  method: status
                  params: {}
                  id: "2"
              Shutdown with typed params:
                description: Shutdown with typed params
                value:
                  method: shutdown
                  params:
                    reason: Maintenance
                    delay: 30
                  id: "3"
        required: true
      responses:
        "200":
          description: RPC call executed successfully with discriminated result
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UnifiedRpcResponse'
  /api/users:
    get:
      tags:
      - User Management
      summary: Get all users
      description: Retrieves a list of all users in the system
      operationId: getAllUsers
      responses:
        "200":
          description: Successfully retrieved all users
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponse'
    post:
      tags:
      - User Management
      summary: Create a new user
      description: Creates a new user with the provided information
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
        required: true
      responses:
        "201":
          description: User created successfully
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponse'
        "400":
          description: Invalid input data
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ApiResponse'
components:
  schemas:
    UpdateUserRequest:
      type: object
      properties:
        email:
          type: string
          description: Updated email address (optional)
          example: john.updated@example.com
        fullName:
          maxLength: 100
          minLength: 0
          type: string
          description: Updated full name (optional)
          example: John Updated Doe
      description: Request payload for updating an existing user
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          type: object
    RpcError:
      type: object
      properties:
        code:
          type: integer
          description: Error code
          format: int32
          example: -1
        message:
          type: string
          description: Error message
          example: "Unknown method: invalid_method"
      description: RPC error details
    RebootRequest:
      type: object
      properties:
        reason:
          type: string
          description: Reason for reboot
          example: Scheduled maintenance
        force:
          type: boolean
          description: Force reboot without graceful shutdown
          example: false
      description: Reboot request parameters
    RestartRequest:
      type: object
      properties:
        service:
          type: string
          description: Service to restart
          example: system
      description: Restart request parameters
    ShutdownRequest:
      type: object
      properties:
        reason:
          type: string
          description: Reason for shutdown
          example: Planned maintenance
        delay:
          type: integer
          description: Delay in seconds before shutdown
          format: int32
          example: 30
      description: Shutdown request parameters
    StatusRequest:
      type: object
      description: Status request parameters (empty)
    UnifiedRpcRequest:
      required:
      - id
      - method
      type: object
      properties:
        rebootParams:
          $ref: '#/components/schemas/RebootRequest'
        statusParams:
          $ref: '#/components/schemas/StatusRequest'
        shutdownParams:
          $ref: '#/components/schemas/ShutdownRequest'
        restartParams:
          $ref: '#/components/schemas/RestartRequest'
        versionParams:
          $ref: '#/components/schemas/VersionRequest'
        method:
          type: string
          description: RPC method name
          example: reboot
          enum:
          - reboot
          - status
          - shutdown
          - restart
          - version
        params:
          type: object
          description: Method-specific parameters discriminated by method field
          oneOf:
          - $ref: '#/components/schemas/RebootRequest'
          - $ref: '#/components/schemas/StatusRequest'
          - $ref: '#/components/schemas/ShutdownRequest'
          - $ref: '#/components/schemas/RestartRequest'
          - $ref: '#/components/schemas/VersionRequest'
        id:
          type: string
          description: Request identifier
          example: "1"
      description: Unified RPC request with discriminated params field based on method
    VersionRequest:
      type: object
      description: Version request parameters (empty)
    RebootResponse:
      type: object
      properties:
        status:
          type: string
          description: Operation status
          example: reboot_initiated
        reason:
          type: string
          description: Reason for reboot
          example: Scheduled maintenance
        force:
          type: boolean
          description: Whether force reboot was used
          example: false
        timestamp:
          type: string
          description: Timestamp of the operation
          example: 2023-11-27T10:30:00
        estimatedDowntime:
          type: string
          description: Estimated downtime
          example: 30 seconds
      description: Reboot operation result
    RestartResponse:
      type: object
      properties:
        status:
          type: string
          description: Operation status
          example: restart_initiated
        service:
          type: string
          description: Service being restarted
          example: system
        timestamp:
          type: string
          description: Timestamp of the operation
          example: 2023-11-27T10:30:00
        estimatedDuration:
          type: string
          description: Estimated duration for restart
          example: 10 seconds
      description: Restart operation result
    ShutdownResponse:
      type: object
      properties:
        status:
          type: string
          description: Operation status
          example: shutdown_scheduled
        reason:
          type: string
          description: Reason for shutdown
          example: Planned maintenance
        delaySeconds:
          type: integer
          description: Delay in seconds
          format: int32
          example: 30
        timestamp:
          type: string
          description: Timestamp of the operation
          example: 2023-11-27T10:30:00
      description: Shutdown operation result
    StatusResponse:
      type: object
      properties:
        uptime:
          type: string
          description: System uptime
          example: "2 days, 14 hours, 23 minutes"
        memoryUsage:
          type: string
          description: Memory usage percentage
          example: 65%
        cpuUsage:
          type: string
          description: CPU usage percentage
          example: 12%
        diskUsage:
          type: string
          description: Disk usage percentage
          example: 78%
        activeConnections:
          type: integer
          description: Number of active connections
          format: int32
          example: 42
        timestamp:
          type: string
          description: Timestamp of the status check
          example: 2023-11-27T10:30:00
      description: System status information
    UnifiedRpcResponse:
      type: object
      properties:
        rebootResult:
          $ref: '#/components/schemas/RebootResponse'
        statusResult:
          $ref: '#/components/schemas/StatusResponse'
        shutdownResult:
          $ref: '#/components/schemas/ShutdownResponse'
        restartResult:
          $ref: '#/components/schemas/RestartResponse'
        versionResult:
          $ref: '#/components/schemas/VersionResponse'
        result:
          type: object
          description: Method-specific result discriminated by method field
          oneOf:
          - $ref: '#/components/schemas/RebootResponse'
          - $ref: '#/components/schemas/StatusResponse'
          - $ref: '#/components/schemas/ShutdownResponse'
          - $ref: '#/components/schemas/RestartResponse'
          - $ref: '#/components/schemas/VersionResponse'
        error:
          $ref: '#/components/schemas/RpcError'
        id:
          type: string
          description: Request identifier
          example: "1"
        method:
          type: string
          description: RPC method name for discriminator
          example: reboot
      description: Unified RPC response with discriminated result field based on method
    VersionResponse:
      type: object
      properties:
        application:
          type: string
          description: Application name
          example: RPC-REST Service
        version:
          type: string
          description: Application version
          example: 1.0.0
        build:
          type: string
          description: Build number
          example: "20241127"
        javaVersion:
          type: string
          description: Java version
          example: 17.0.1
        springBootVersion:
          type: string
          description: Spring Boot version
          example: 3.2.0
      description: Application version information
    CreateUserRequest:
      required:
      - email
      - fullName
      - username
      type: object
      properties:
        username:
          maxLength: 50
          minLength: 3
          type: string
          description: Username for the user
          example: johndoe
        email:
          type: string
          description: Email address of the user
          example: john.doe@example.com
        fullName:
          maxLength: 100
          minLength: 0
          type: string
          description: Full name of the user
          example: John Doe
      description: Request payload for creating a new user